Read CarReport backup and send via MQTT to HA
I keep track of my expenses (fuel and other) for my car via CarReport, and
configured it to make an automatic backup. The backup is located in a folder
synced via SyncThing. The backups are SQLite. This code then reads the backup,
and for every car in the backup, sends a number of metrics to a MQTT instance.
Home Assistant is then configured to read this MQTT and parse the metrics.

- CarReport: https://f-droid.org/en/packages/me.kuehle.carreport/
- Home Assistant: home-assistant.io/

Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>